// src/app/interfaces/drinks.ts
export interface Drinks {
    drinks:any;
    strDrink: string;
    strDrinkThumb: string;
    idDrink: string;
    price: string|number;
    strCategory: string;
    strInstructions: string;
    strIngredient1: string | null;
    strIngredient2: string | null;
    strIngredient3: string | null;
}